rename forcedaction to forced-actions
