+++ packages/frontend/src/scripts/forced-actions/forcedActionForm.ts
@@ -1,5 +1,5 @@
-import { FormId } from '../../view/components/forcedaction/form/ids'
-import { ForcedActionFormProps } from '../../view/pages/forcedactions/ForcedActionFormProps'
+import { FormId } from '../../view/components/forced-actions/form/ids'
+import { ForcedActionFormProps } from '../../view/pages/forced-actions/ForcedActionFormProps'
 import { FormElements, getFormElements } from './getFormElements'
 import { getInitialState, nextFormState } from './state'
 import { submit } from './submit'
