+++ packages/frontend/src/view/components/forced-actions/ForcedActionCard.tsx
