+++ packages/frontend/src/view/components/forced-actions/form/AmountInput.tsx
@@ -1,27 +1,26 @@
-import { AssetId } from '@explorer/types'
 import React from 'react'
 
 import { assetToInfo } from '../../../../utils/assetUtils'
 import { formatCurrencyInput } from '../../../../utils/formatUtils'
-import { ForcedActionFormProps } from '../../../pages/forcedactions/ForcedActionFormProps'
+import { ForcedActionFormProps } from '../../../pages/forced-actions/ForcedActionFormProps'
 import { AssetWithLogo } from '../../common/AssetWithLogo'
 import { FormId } from './ids'
 
-export function PriceInput(props: ForcedActionFormProps) {
-  const usdcInfo = assetToInfo(AssetId.USDC)
-  const assetDetails = props.assets.find(
+export function AmountInput(props: ForcedActionFormProps) {
+  const assetInfo = assetToInfo(props.selectedAsset)
+  const balance = props.assets.find(
     (asset) => asset.assetId === props.selectedAsset
-  )
-  const price = (assetDetails?.priceUSDCents ?? 0n) * 10000n
-  const priceFormatted = formatCurrencyInput(price, AssetId.USDC)
+  )?.balance
+  const sign = balance && balance < 0 ? '-' : ''
+  const formattedBalance = formatCurrencyInput(balance, props.selectedAsset)
 
   return (
     <div className="flex gap-2">
       <div className="flex flex-1 flex-col">
-        <span className="text-sm text-zinc-500">Price</span>
+        <span className="text-sm text-zinc-500">Amount</span>
         <div>
           <input
-            id={FormId.PriceInput}
+            id={FormId.AmountInput}
             type="text"
             autoComplete="off"
             placeholder="0.00"
@@ -30,16 +29,17 @@ export function PriceInput(props: ForcedActionFormProps) {
         </div>
         <div
           id={FormId.AmountErrorView}
-          className="hidden font-medium text-red-500"
+          className="hidden text-sm font-medium text-red-500"
         >
           Amount too large
         </div>
       </div>
       <div className="flex flex-col items-end">
         <span className="text-sm text-zinc-500">
-          Market price: {priceFormatted}
+          Balance: {sign}
+          {formattedBalance}
         </span>
-        <AssetWithLogo assetInfo={usdcInfo} />
+        <AssetWithLogo assetInfo={assetInfo} />
       </div>
     </div>
   )
