+++ packages/frontend/src/view/components/forced-actions/form/TotalInput.tsx
