+++ packages/frontend/src/view/index.ts
@@ -7,7 +7,7 @@ export * from './old/positions'
 export * from './old/state-updates'
 export * from './old/transaction-form'
 /* eslint-enable no-restricted-imports */
-export * from './pages/forcedactions/ForcedTradePage'
-export * from './pages/forcedactions/ForcedWithdrawPage'
+export * from './pages/forced-actions/ForcedTradePage'
+export * from './pages/forced-actions/ForcedWithdrawPage'
 export * from './pages/HomePage'
 export * from './pages/UserPage'
