+++ packages/frontend/src/view/pages/forced-actions/ForcedTradePage.tsx
@@ -6,11 +6,11 @@ import { Card } from '../../components/common/Card'
 import { Link } from '../../components/common/Link'
 import { OrderedList } from '../../components/common/OrderedList'
 import { Page } from '../../components/common/page/Page'
-import { ForcedActionCard } from '../../components/forcedaction/ForcedActionCard'
-import { AmountInput } from '../../components/forcedaction/form/AmountInput'
-import { FormId } from '../../components/forcedaction/form/ids'
-import { PriceInput } from '../../components/forcedaction/form/PriceInput'
-import { TotalInput } from '../../components/forcedaction/form/TotalInput'
+import { ForcedActionCard } from '../../components/forced-actions/ForcedActionCard'
+import { AmountInput } from '../../components/forced-actions/form/AmountInput'
+import { FormId } from '../../components/forced-actions/form/ids'
+import { PriceInput } from '../../components/forced-actions/form/PriceInput'
+import { TotalInput } from '../../components/forced-actions/form/TotalInput'
 import { reactToHtml } from '../../reactToHtml'
 import {
   ForcedActionFormProps,
